youtube_live_streamer.py
import streamlit as st
import cv2
import numpy as np
from datetime import datetime
import json
import os

class YouTubeLiveStreamer:
    def __init__(self):
        self.stream_key = None
        self.camera = None
        self.is_streaming = False
        
    def setup_camera(self):
        """카메라 설정"""
        self.camera = cv2.VideoCapture(0)
        if not self.camera.isOpened():
            st.error("카메라를 열 수 없습니다.")
            return False
        return True
    
    def start_stream(self, stream_key):
        """라이브 스트리밍 시작"""
        self.stream_key = stream_key
        self.is_streaming = True
        st.success("라이브 스트리밍이 시작되었습니다!")
        
    def stop_stream(self):
        """라이브 스트리밍 중지"""
        self.is_streaming = False
        if self.camera:
            self.camera.release()
        st.info("라이브 스트리밍이 중지되었습니다.")
        
    def get_stream_info(self):
        """스트리밍 정보 반환"""
        return {
            "timestamp": datetime.now().isoformat(),
            "status": "streaming" if self.is_streaming else "stopped",
            "stream_key": self.stream_key
        }

def main():
    st.title("🎥 YouTube Live Streamer")
    st.markdown("### boneekim의 YouTube 라이브 스트리밍 도구")
    
    streamer = YouTubeLiveStreamer()
    
    # 사이드바 설정
    with st.sidebar:
        st.header("스트리밍 설정")
        stream_key = st.text_input("YouTube 스트림 키", type="password")
        
        if st.button("카메라 설정"):
            if streamer.setup_camera():
                st.success("카메라 설정 완료!")
        
        if st.button("스트리밍 시작"):
            if stream_key:
                streamer.start_stream(stream_key)
            else:
                st.error("스트림 키를 입력해주세요.")
                
        if st.button("스트리밍 중지"):
            streamer.stop_stream()
    
    # 메인 영역
    col1, col2 = st.columns([2, 1])
    
    with col1:
        st.subheader("라이브 프리뷰")
        video_placeholder = st.empty()
        
        # 비디오 피드 표시 (실제 구현에서는 실시간 스트리밍 로직 추가)
        if streamer.is_streaming:
            st.info("🔴 LIVE - 스트리밍 중...")
        else:
            st.info("📹 스트리밍 대기 중...")
    
    with col2:
        st.subheader("스트리밍 정보")
        info = streamer.get_stream_info()
        st.json(info)
        
        st.subheader("실시간 통계")
        st.metric("시청자 수", "0")
        st.metric("좋아요 수", "0")
        st.metric("댓글 수", "0")

if __name__ == "__main__":
    main()
