requirements.txt
streamlit==1.28.0
opencv-python==4.8.0.74
numpy==1.24.3
python-dotenv==1.0.0
google-api-python-client==2.95.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
pytube==15.0.0
pillow==10.0.0
pandas==2.0.3
matplotlib==3.7.1
seaborn==0.12.2
requests==2.31.0
